move _magicgui to utils, add show_result callback
