+++ napari/__init__.py
@@ -48,7 +48,7 @@ from .view_layers import (
 )
 from ._qt import gui_qt
 from ._version import get_versions
-from .utils import sys_info
+from .utils import sys_info, _magicgui
 
 __version__ = get_versions()['version']
 del get_versions
@@ -59,8 +59,5 @@ del get_versions
 # see: https://github.com/napari/napari/issues/925
 from scipy import stats  # noqa: F401
 
-del stats
-
-from . import _magicgui
-
 del _magicgui
+del stats
